refactor: header componentization
