+++ src/components/Header/Logo.tsx
@@ -1,0 +1,12 @@
+import { Text } from "@chakra-ui/react";
+
+export function Logo() {
+  return (
+    <Text fontSize="3xl" fontWeight="bold" letterSpacing="tight" w="64">
+      dashgo
+      <Text as="span" color="pink.500">
+        .
+      </Text>
+    </Text>
+  );
+}
