+++ src/pages/dashboard.tsx
@@ -84,6 +84,7 @@ export default function Dashboard() {
             <Text fontSize="lg" mb="4">
               Taxa de abertura
             </Text>
+            <Chart options={options} series={series} type="area" height={160} />
           </Box>
         </SimpleGrid>
       </Flex>
